+++ Together/Together/Models/Completions/CompletionChoicesChunk.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using Together.Models.Common;
 
 namespace Together.Models.Completions;
@@ -6,7 +7,7 @@ public class CompletionChoicesChunk
 {
     public int? Index { get; set; }
     public float? Logprobs { get; set; }
-    public int? Seed { get; set; }
+    public ulong? Seed   { get; set; }
     public FinishReason? FinishReason { get; set; }
     public DeltaContent Delta { get; set; }
 }
