+++ Together/Together/Models/Images/ImageRequest.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace Together.Models.Images;
 
 public class ImageRequest
@@ -5,7 +7,7 @@ public class ImageRequest
     public string Prompt { get; set; }
     public string Model { get; set; }
     public int? Steps { get; set; } = 20;
-    public int? Seed { get; set; }
+    public ulong? Seed   { get; set; }
     public int? N { get; set; } = 1;
     public int? Height { get; set; } = 1024;
     public int? Width { get; set; } = 1024;
